isa dma: Use size_t to transfer remaining buffer size.
Be more careful with dma size bit magic.

+++ uspace/drv/bus/isa/i8237.h
@@ -38,7 +38,7 @@
 #define DRV_BUS_ISA_I8237_H
 
 extern int dma_channel_setup(unsigned, uint32_t, uint16_t, uint8_t);
-extern int dma_channel_remain(unsigned, uint16_t *);
+extern int dma_channel_remain(unsigned, size_t *);
 
 #endif
